rewrite ... not allowed if federated in
