+++ src/test/java/org/apache/sysds/test/component/resource/RecompilationTest.java
@@ -235,7 +235,7 @@ public class RecompilationTest extends AutomatedTestBase {
 
 		// original compilation used for comparison
 		Program expectedProgram = ResourceCompiler.compile(HOME+"mm_test.dml", nvargs);
-		Program recompiledProgram = runTest(precompiledProgram, expectedProgram, driverMemory, numberExecutors, executorMemory);
+		Program recompiledProgram = runTest(precompiledProgram, expectedProgram, driverMemory, numberExecutors, executorMemory, new StringBuilder());
 
 		Optional<Instruction> mmInstruction = ((BasicProgramBlock) recompiledProgram.getProgramBlocks().get(0)).getInstructions().stream()
 				.filter(inst -> (Objects.equals(expectedSparkExecType, inst instanceof SPInstruction) && Objects.equals(inst.getOpcode(), expectedOpcode)))
@@ -257,7 +257,7 @@ public class RecompilationTest extends AutomatedTestBase {
 		}
 		// original compilation used for comparison
 		Program expectedProgram = ResourceCompiler.compile(HOME+"mm_transpose_test.dml", nvargs);
-		Program recompiledProgram = runTest(precompiledProgram, expectedProgram, driverMemory, numberExecutors, executorMemory);
+		Program recompiledProgram = runTest(precompiledProgram, expectedProgram, driverMemory, numberExecutors, executorMemory, new StringBuilder());
 		Optional<Instruction> mmInstruction = ((BasicProgramBlock) recompiledProgram.getProgramBlocks().get(0)).getInstructions().stream()
 				.filter(inst -> (Objects.equals(expectedSparkExecType, inst instanceof SPInstruction) && Objects.equals(inst.getOpcode(), expectedOpcode)))
 				.findFirst();
@@ -273,8 +273,9 @@ public class RecompilationTest extends AutomatedTestBase {
 								  Map<String, String> nvargs) throws IOException {
 		// pre-compiled program using default values to be used as source for the recompilation
 		Program precompiledProgram = generateInitialProgram(HOME+dmlScript, nvargs);
-		System.out.println("precompiled");
-		System.out.println(Explain.explain(precompiledProgram));
+		StringBuilder sb = new StringBuilder();
+		sb.append("\n\nprecompiled\n");
+		sb.append(Explain.explain(precompiledProgram));
 		if (numberExecutors > 0) {
 			ResourceCompiler.setSparkClusterResourceConfigs(driverMemory, driverThreads, numberExecutors, executorMemory, executorThreads);
 		} else {
@@ -282,13 +283,13 @@ public class RecompilationTest extends AutomatedTestBase {
 		}
 		// original compilation used for comparison
 		Program expectedProgram = ResourceCompiler.compile(HOME+dmlScript, nvargs);
-		System.out.println("expected");
-		System.out.println(Explain.explain(expectedProgram));
-		runTest(precompiledProgram, expectedProgram, driverMemory, numberExecutors, executorMemory);
+		sb.append("\n\nexpected\n");
+		sb.append(Explain.explain(expectedProgram));
+		runTest(precompiledProgram, expectedProgram, driverMemory, numberExecutors, executorMemory, sb);
 	}
 
-	private Program runTest(Program precompiledProgram, Program expectedProgram, long driverMemory, int numberExecutors, long executorMemory) {
-		if (DEBUG_MODE) System.out.println(Explain.explain(expectedProgram));
+	private Program runTest(Program precompiledProgram, Program expectedProgram, long driverMemory, int numberExecutors, long executorMemory, StringBuilder sb) {
+		if (DEBUG_MODE) sb.append(Explain.explain(expectedProgram));
 		Program recompiledProgram;
 		if (numberExecutors == 0) {
 			ResourceCompiler.setSingleNodeResourceConfigs(driverMemory, driverThreads);
@@ -303,19 +304,19 @@ public class RecompilationTest extends AutomatedTestBase {
 			);
 			recompiledProgram = ResourceCompiler.doFullRecompilation(precompiledProgram);
 		}
-		System.out.println("recompiled");
-		System.out.println(Explain.explain(recompiledProgram));
+		sb.append("\n\nrecompiled\n");
+		sb.append(Explain.explain(recompiledProgram));
 
-		if (DEBUG_MODE) System.out.println(Explain.explain(recompiledProgram));
-		assertEqualPrograms(expectedProgram, recompiledProgram);
+		if (DEBUG_MODE) sb.append(Explain.explain(recompiledProgram));
+		assertEqualPrograms(expectedProgram, recompiledProgram, sb);
 		return recompiledProgram;
 	}
 
-	private void assertEqualPrograms(Program expected, Program actual) {
+	private void assertEqualPrograms(Program expected, Program actual, StringBuilder sb) {
 		// strip empty blocks basic program blocks
 		String expectedProgramExplained = stripGeneralAndReplaceRandoms(Explain.explain(expected));
 		String actualProgramExplained = stripGeneralAndReplaceRandoms(Explain.explain(actual));
-		Assert.assertEquals(expectedProgramExplained, actualProgramExplained);
+		Assert.assertEquals(sb.toString(), expectedProgramExplained, actualProgramExplained);
 	}
 
 	private String stripGeneralAndReplaceRandoms(String explainedProgram) {
